initial: introduce μeb-archive tool
link rotting is a problem, and this is my attempt to solve it.

the initial assumptions were:
(a) i can use firefox bookmarks to transfer information about links i
want to archive for the offline usage, say, store there bookmarks in a
folder with pre-defined name ("archive", "offline", etc). and
(b) it is possible to inspect firefox internals in a machine-friendly
way, and extract these bookmarks known the folder name.