feat: introduce basic web-page archiving tool
this commit implements a simple solution, based on firefox bookmarks.
the approach is the following:

1. look for bookmarks database in the firefox's profile,
   by inspecting the content of ~/.mozilla/firefox/ directory.

2. find all bookmarks in the pre-defined folder, dedicated for the
   archiving. take a look at the -folder flag.

3. download web-pages and its related resources (.js, .css, images),
   using wget in page-requisites mode. then save it in the specified
   location (see -archive flag), keeping the website directory
   structure inside the archive root.

4. build an index page on top of the archive.

please note that ueb-archive tool does not store any metadata about what
and when and how successful were downloaded. thus, on each run it will
fetch everything it can get. the naive approach is to run the tool by
cron every 10 days or so.

nice-to-haves that will probably be implemented soon:
- better index page: more meta-data, some css, ordering;
- check atime/mtime, do not download recently updated pages;
- store archive meta-data in a machine-friendly format (sqlite?
  json-mutex-db??);
- hooks for notifications (just run a specified script when done);